Code cleanup ch07

+++ ch07/Transforms/Transforms/FlowerTransformView.m
@@ -2,8 +2,27 @@
 //  FlowerTransformView.m
 //  Transforms
 //
-//  Created by Rob Napier on 6/29/11.
-//  Copyright 2011 __MyCompanyName__. All rights reserved.
+//  Copyright (c) 2012 Rob Napier
+//
+//  This code is licensed under the MIT License:
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a
+//  copy of this software and associated documentation files (the "Software"),
+//  to deal in the Software without restriction, including without limitation
+//  the rights to use, copy, modify, merge, publish, distribute, sublicense,
+//  and/or sell copies of the Software, and to permit persons to whom the
+//  Software is furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+//  DEALINGS IN THE SOFTWARE.
 //
 
 #import "FlowerTransformView.h"
@@ -29,26 +48,26 @@ CGAffineTransformMakeScaleTranslate(CGFloat sx, CGFloat sy,
   UIBezierPath *path = [UIBezierPath bezierPath];
   [path addArcWithCenter:CGPointMake(0, -1)
                   radius:1
-              startAngle:-M_PI
+              startAngle:(CGFloat)-M_PI
                 endAngle:0
                clockwise:YES];
   [path addArcWithCenter:CGPointMake(1, 0)
                   radius:1
-              startAngle:-M_PI_2
-                endAngle:M_PI_2
+              startAngle:(CGFloat)-M_PI_2
+                endAngle:(CGFloat)M_PI_2
                clockwise:YES];
   [path addArcWithCenter:CGPointMake(0, 1)
                   radius:1
               startAngle:0
-                endAngle:M_PI
+                endAngle:(CGFloat)M_PI
                clockwise:YES];
   [path addArcWithCenter:CGPointMake(-1, 0)
                   radius:1
-              startAngle:M_PI_2
-                endAngle:-M_PI_2
+              startAngle:(CGFloat)M_PI_2
+                endAngle:(CGFloat)-M_PI_2
                clockwise:YES];
   
-  CGFloat scale = floor((MIN(size.height, size.width)
+  CGFloat scale = floorf((MIN(size.height, size.width)
                          - margin) / 4);
   
   CGAffineTransform transform;
